309
git-svn-id: http://caml.inria.fr/svn/ocaml/branches/jocamltrunk@7186 f963ae5c-01c2-4b8c-9fe0-0dff7051ff02

+++ otherlibs/win32unix/lockf.c
@@ -62,7 +62,7 @@ static void set_file_pointer(HANDLE h, LARGE_INTEGER dest,
   LONG high = dest.HighPart;
   DWORD ret = SetFilePointer(h, dest.LowPart, &high, method);
   if (ret == INVALID_SET_FILE_POINTER) {
-    long err = GetLastError();
+    DWORD err = GetLastError();
     if (err != NO_ERROR) { win32_maperr(err); uerror("lockf", Nothing); }
   }
   if (cur != NULL) { cur->LowPart = ret; cur->HighPart = high; }
